<Hp, Death and Status System>
Added the Hp which gives the player health, and all the necessary methods for others to interface with the player's hp.

Added the Death system which locks the players controls when they run out of hp and then triggers any OnDeathListeners subscribed to the player.

Added Spike trap and checkpoint gameobjects to demonstrate the HP and OnDeath systems. Spike Traps extend on a brief delay when touched an deal damage on contact while extended. The last checkpoint touched will call Player_Logic.respawn() from it's OnDeath() method when the player dies.

Added the Status System. This system allows for the assigning of id-identified strings to a list of Statuses on the player character. On their own they do nothing but can be used by external scripts for assigning data to the player.

Added an electricfied lever, water patch and ExampleStatusEffectMonitor as a means of demostrating the status system by making the water patch apply the "Wet" status with a value of 10 to the player. The monitor then decrements the value of "Wet" over time before removing it once it reaches zero. If they player attempts to use the electrified lever while they have the "Wet" status they take damage.