<Massive Refactor>
Refactored the whole project, renaming files and reorganising folders to properly organise assets for the game jam and to align with the new style guide rules.

Reworked the Layers and Sorting Layers to the versions intended to be used in the Game Jam.

Added Rules.txt detailing the current version of the rules for the game jam. Also includes style guide rules.