<Revised Comments>
Revised a lot of the comments and added a lot of omitted comments to parts of the Essential Prefabs core of the project.

For those looking for the original unedited comments from the time of the original game jam, you can find them in the <Original-Assembled-Version> branch of the repository.